+++ packages/plugin-compiler/src/test/kotlin/io/realm/kotlin/compiler/GenerationExtensionTest.kt
@@ -254,6 +254,11 @@ class GenerationExtensionTest {
             // Linking objects
             "linkingObjectsByList" to PropertyType.RLM_PROPERTY_TYPE_LINKING_OBJECTS,
             "linkingObjectsBySet" to PropertyType.RLM_PROPERTY_TYPE_LINKING_OBJECTS,
+
+            // @PersistedName annotated fields
+            "persistedNameStringField" to PropertyType.RLM_PROPERTY_TYPE_STRING,
+            "persistedNameChildField" to PropertyType.RLM_PROPERTY_TYPE_OBJECT,
+            "persistedNameLinkingObjectsField" to PropertyType.RLM_PROPERTY_TYPE_LINKING_OBJECTS
         )
         assertEquals(expectedProperties.size, properties.size)
         properties.map { property ->