+++ packages/test-base/src/androidAndroidTest/kotlin/io/realm/kotlin/test/shared/QueryTests.kt
@@ -47,6 +47,7 @@ import io.realm.kotlin.types.RealmInstant
 import io.realm.kotlin.types.RealmList
 import io.realm.kotlin.types.RealmObject
 import io.realm.kotlin.types.RealmUUID
+import io.realm.kotlin.types.annotations.PersistedName
 import kotlinx.coroutines.TimeoutCancellationException
 import kotlinx.coroutines.async
 import kotlinx.coroutines.channels.Channel
@@ -2572,6 +2573,9 @@ class QuerySample() : RealmObject {
         this.intField = intField
         this.stringField = stringField
     }
+
+    @PersistedName("persistedNameStringField")
+    var publicNameStringField: String = "Realm"
 
     var stringField: String = "Realm"
     var byteField: Byte = 0